use stable sort for choices
choices_search_worker sorts choices in assumption that corresponding
pointers are contiguously allocated. If this is not the case, the the
order will be broken (that's because libc qsort is not guaranteed to be
stable).

This commit solves this problem by using hand-written merge sort
algorithm, which is always stable.